backend/requirements/requirements-focal.txt
acme==1.1.0
automat==0.8.0
chardet==3.0.4
cffi==1.14.0
cryptography==2.8
h2==3.1.1
idna==2.8
priority==1.3.0
pyopenSSL==19.0.0
pyasn1==0.4.2
pyasn1_modules==0.2.1
pynacl==1.3.0
python_debian==0.1.36
python_gnupg==0.4.5
python-dotenv==0.17.0
requests==2.22.0
service_identity==18.1.0
simple-salesforce==1.11.1
sqlalchemy==1.3.12
twisted==18.9.0
txtorcon==18.3.0
zope.interface==4.7.1
